backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-ai
python-socketio
sqlalchemy
psycopg2-binary
openbb
pybind11
python-dotenv
httpx
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
openai
requests
asyncpg
logfire[fastapi]
